About

DIVACoast.jl is a julia library for coastal impact and adaptation modelling. The library provides data types and algorithms to script assessment runs for different coastal impact and adaptation research questions. DIVACoast.jl is provided by the Global Climate Forum via GitLab.

Download & Installation

Ensure you have Julia installed on your system. You can download Julia from the official julia website.

You can get the latest (unstable) version by cloning the library repository to your machine. Since there is not stable version at the moment you have to switch to the development branch afterwards.

git clone https://gitlab.com/globalclimateforum/DIVACoast.jl.git
git checkout development

The DIVA library has several dependencies with multiple ways to install them:

Install requirements to local environment If you want to install all required packages to your local environment, easiest is to execute the install_packages.jl script within the parent directory of the repository.

Instantiate diva_library project If you want to use the divalibrary project environment you can execute the DIVACoast.jl script in the './src' directory. The script will activate the project and install all (instantiate) dependencies automatically to a environment called divalibrary.

Setting environent variables We recommend to setup environment variables for the DIVA library directory and your data directory. ```

You can include the diva library in your script by:

include(<path_to_diva>/diva_library/src/DIVACoast.jl); using .jdiva

Data structures

Impact Model

Main.DIVACoast.ComposedImpactModel — Type

ComposedImpactModel{ID_TYPE1,ID_TYPE2,DATA,CIU} <: CoastalImpactUnit where {CIU <: CoastalImpactUnit}

A ComposedImpactModel combines multiple LocalCoastalImpactModels into a combined datastructure.

Main.DIVACoast.LocalCoastalImpactModel — Type

LocalCoastalImpactModel{DT<:Real, IDT, DATA} <: CoastalImpactUnit

A LocalCoastalImpactModel combines a surge model (distribution) and a coastal plain model (HypsometricProfile). It also holds the current protection level.

Hypsometric Profile

Main.DIVACoast.HypsometricProfile — Type

HypsometricProfile(width::DT, width_unit::String,
elevations::Array{DT}, elevation_unit::String, area::Array{DT}, area_unit::String,
s_exposure::StructArray{T1}, s_exposure_units::Array{String},
d_exposure::StructArray{T2}, d_exposure_units::Array{String}) where {DT<:Real,T1,T2}

A HypsometricProfile represents the variation in elevation from the coastline to inland areas. It can be constructed manually or by using load_hsps_nc() and a NetCDF-file.

Main.DIVACoast.load_hsps_nc — Function

load_hsps_nc(::Type{IT}, ::Type{DT}, filename::String)::Dict{IT,HypsometricProfile{DT}} where {IT<:Integer,DT<:Real}

Load a netcdf file into hypsometric profiles. IT is the indextype (an integer type which is used to adress a specific profile), DT is the datatype (a floating point type which is used to store the data internally). Hypsometric profiles are not compressed.

Examples

julia> test = load_hsps_nc(Int32, Float32, "test.nc")
Dict{Int32, HypsometricProfile{Float32}} with 2716 entries:
  2108 => HypsometricProfile{Float32}(1.0, Float32[-5.0, -4.9, -4.8, -4.7, -4.6, -4.5, -4.4, -4.3, -4.2, -4.1  …  19.1, 19.2, 19.3, 19.4, 19.5, 19.6, 19.7, 19.8, 19.9, 20.0], Float32[0.0, 0.0, 0.0, 0.0, …

Main.DIVACoast.to_hypsometric_profile — Function

tohypsometricprofile(sga::SparseGeoArray{DT,IT}, width::DT2, minelevation::DT2, maxelevation::DT2, elevation_incr::DT2)::HypsometricProfile where {DT<:Real,IT<:Integer,DT2<:Real}

A hypsometric profile is constructed from an elevation geotiff representing elevations.

Base.:+ — Function

Addtion of two Hypsometric Profiles. Adds (combines) the folling properties of the HypsometricProfiles:

Elevation: Combine Increments

width: Adds the width of both HypsometricProfiles

cummulativeArea: Adds the cummulative are of both HypsometricProfiles

static Exposure: Adds the cummulative static exposure of both HypsometricProfiles

dynamic Exposure: Adds the dynamic exposure of both HypsometricProfiles

NOTE: Distances are not added, and need to be recalculated!

Adapt

Main.DIVACoast.add_static_exposure! — Function

Adds static exposure to a HypsometricProfile.

Main.DIVACoast.add_dynamic_exposure! — Function

Adds dynamic exposure to a HypsometricProfile.

Main.DIVACoast.remove_static_exposure! — Function

Removes static exposure from a HypsometricProfile

Main.DIVACoast.remove_dynamic_exposure! — Function

Removes dynamic exposure from a HypsometricProfile

Main.DIVACoast.sed! — Function

Applies factor to exposed assets / population. (socio-economic development)

Main.DIVACoast.sed_above! — Function

Applies socio-economic development (factor) above a certain elevation.

Main.DIVACoast.sed_below! — Function

Applies socio-economic development (factor) below a certain elevation.

Main.DIVACoast.remove_below! — Function

Removes exposed assets / population below a certain elevation.

Main.DIVACoast.add_above! — Function

Adds assets / population above a certain elevation.

Main.DIVACoast.add_between! — Function

Adds assets / population between certain elevations.

Main.DIVACoast.compress! — Function

compress!(hspf::HypsometricProfile)

Comress a hypsometric profile by removing colinear points. Calculations on compressed hypsometric profiles can be faster. Idempotent operation.

Analysis

Calculate Damages

Main.DIVACoast.expected_damage_bathtub_standard_ddf — Function

expecteddamagebathtubstandardddf(LocalCoastalModel::LocalCoastalModel{DT}, hddarea::DT, hddsstatic::Array{DT}, hdds_dynamic::Array{DT})

This function calculates the annual expected damage for one local coastal model (Hypsometric Profile and Extreme surge distribution) by integrating the product of damages and the pdf (probability disctribution function) of the surge model over all possible extreme values. The output are annual expected damage for area, static and dynamic. The standard depth damage function is used to estimate flood damages.

Main.DIVACoast.expected_damage_bathtub — Function

expecteddamagebathtub(LocalCoastalModel::LocalCoastalModel{DT}, ddfarea::Function, ddfstatic::Array{Function}, ddf_dynamic::Array{Function})

This function calculates the annual expected damage for one local coastal model (Hypsometric Profile and Extreme surge distribution) by integrating the product of damages and the pdf (probability disctribution function) of the surge model. The output are annual expected damage for area, static and dynamic. The depth damage functions inserted as inputs are used in this functino to calculate flood damages.

Get Exposure

Main.DIVACoast.exposure_below_bathtub — Function

exposure_below_bathtub(hspf::HypsometricProfile{DT}, e::Real) where {DT<:Real}

Calculate the cumulative area, static exposure, and dynamic exposure below elevation (e) for a hypsometric profile. The function handles different cases based on the elevation's presence in the profile and its position.

Arguments

hspf::HypsometricProfile{DT}: The hypsometric profile with elevation, area and exposure data. e::Real: The elevation threshold for which exposure is calculated (everything underneath this elevation).

Returns

Exposed area, static and dynamic exposure for elevations smaller than e.

Main.DIVACoast.exposure_below_attenuated — Function

exposure_below_attenuated(hspf::HypsometricProfile{DT}, e::Real, att_rates :: Array{<:Real}) where {DT<:Real}
exposure_below_attenuated(hspf::HypsometricProfile{DT}, e::Real, att_rate :: Real) where {DT<:Real}

Calculate the cumulative area, static exposure, and dynamic exposure below elevation (e) for a hypsometric profile, taking into account attenuation.

Arguments

hspf::HypsometricProfile{DT}: The hypsometric profile with elevation, area and exposure data. e::Real: The elevation threshold for which exposure is calculated (everything underneath this elevation). att_rates :: Array{<:Real}: an array of attenuation rates, each given in m/km. A value of 0.3 stands for an attenuation of 0.3m per km of flood extend.att_rate:: Real: an attenuation rate, given in m/km.

If attenuation rates are given as array, the array dimension has to much the array dimension of the elevation array of hspf.

Returns

Exposed area, static and dynamic exposure for elevations smaller than e`, taking into account attenuation.

Main.DIVACoast.attenuate — Function

attenuate(hspf::HypsometricProfile{DT}, wl::Real, att_rates :: Array{<:Real}) where {DT<:Real}
attenuate(hspf::HypsometricProfile{DT}, wl::Real, att_rate :: Real) where {DT<:Real}

Attenuate waterlevel wl for a hypsometric profile.

Arguments

hspf::HypsometricProfile{DT}: The hypsometric profile with elevation, area and exposure data. wl::Real: The elevation threshold for which exposure is calculated (everything underneath this elevation). att_rates :: Array{<:Real}: an array of attenuation rates, each given in m/km. A value of 0.3 stands for an attenuation of 0.3m per km of flood extend.att_rate:: Real: an attenuation rate, given in m/km.

If attenuation rates are given as array, the array dimension has to much the array dimension of the elevation array of hspf.

Returns

The attenuated waterlevel.

Statistics

Main.DIVACoast.estimate_gumbel_distribution — Function

This function tries to fit a gumbel distribution to given data. x is the actual data (e.g. water level). y are the empirical cdf values for the data in x (i.e. values between 0 and 1 - to be interpreted as quantiles). The funtion returns a GeneralizedExtremeValue (GEV) with the third (shape, ξ) parameter being zero. If the cdf fit fails for any reason, the standard gumbel distribution (μ=mean(data), σ=var(data), ξ=0.0) is returned

Main.DIVACoast.estimate_frechet_distribution — Function

This function fits a Frechet Distribution to the inserted data. y should be the return period and x the corresponding water level height. The funtion returns a GeneralizedExtremeValue (GEV).

This function tries to fit a Frechet distribution to given data. x is the actual data (e.g. water level). y are the empirical cdf values for the data in x (i.e. values between 0 and 1 - to be interpreted as quantiles). The funtion returns a GeneralizedExtremeValue (GEV) with the third (shape, ξ) parameter being bigger than zero. If the cdf fit fails for any reason, a standard Frechet distribution (μ=mean(data), σ=var(data), ξ=0.5) is returned

Main.DIVACoast.estimate_gev_distribution — Function

This function fits an extreme value distribution to the inserted data. y should be the return period and x the corresponding water level height. The funtion returns a GeneralizedExtremeValue (GEV) and uses the best fit out of the Gumbel, Frechet and Weibull model based on the summed squared residuals.

Main.DIVACoast.estimate_weibull_distribution — Function

This function fits a Weibull Distribution to the inserted data. y should be the return period and x the corresponding water level height. The funtion returns a GeneralizedExtremeValue (GEV).

This function tries to fit a Weibull distribution to given data. 
    x is the actual data (e.g. water level).
    y are the cempirical df values for the data in x (i.e. values between 0 and 1 - to be interpreted as quantiles). 
The funtion returns a GeneralizedExtremeValue (GEV) with the third (shape, ξ) parameter being smaller than zero. If the cdf fit fails for any reason, 
a standard Weibull distribution (μ=mean(data), σ=var(data), ξ=-0.5) is returned

Main.DIVACoast.estimate_gpd_negative_distribution — Function

This function fits a Generalized Pareto Distribution with negative shape to the inserted data. y should be the return period and x the corresponding water level height. The funtion returns a GeneralizedPareto (GPD).

Main.DIVACoast.estimate_gpd_positive_distribution — Function

This function fits a gpd_positive Distribution to the inserted data. y should be the return period and x the corresponding water level height. The funtion returns a GeneralizedExtremeValue (GEV).

Main.DIVACoast.estimate_gp_distribution — Function

This function fits an extreme value distribution to the inserted data. y should be the return period and x the corresponding water level height. The funtion returns a GeneralizedPareto (GPD) and uses the best fit out of the exponential, positive GPD and negative GPD model based on the summed squared residuals.

Main.DIVACoast.estimate_exponential_distribution — Function

This function tries to fit an exponential distribution to given data. x is the actual data (i.e. water level). y are the cdf values for the data in x (i.e. values between 0 and 1). The funtion returns a GeneralizedPareto (GPD) with the third shape parameter being zero. If the cdf fit fails for any reason, the standard exponential distribution (μ=0) is returned

Main.DIVACoast.plot_comparison_extreme_distributions — Function

plot_comparison_extreme_distributions(x_data::Array{T}, y_data::Array{T})

This function creates a plot that compares six different extreme distribution fits to underlying data. The input is an xdata and ydata array, xdata is the absolute height of extremes and ydata is the annual probability to be below this value. The function fits three different GEV and three different GPD functions to the data and plots all of these. It also highlights the best fit out of these six functions and highlights this in the plot.

Data

Read SLR Data

Main.DIVACoast.SLRScenarioReader — Type

Creates a SLR-Scenario reader around a dataset (NetCDF) using dataset specific variable names for variable (e.g., "SeaLevelRise"), latitude, longitude, time, and quantile After the Wrapper structure was initialized dataset specific functions can be used (e.g, getslrvalue)

Main.DIVACoast.get_slr_value — Function

Gets the Sea Level Rise value at a specific location (lon, lat) in a specific quantile at a specific time.

Main.DIVACoast.get_slr_value_from_cell — Function

Gets the Sea Level Rise value at a specific cell (indexlon, indexlat) in a specific quantile (given by index_qtl) at a specific time. Faster than the previous

Read SSP Data

Main.DIVACoast.SSPWrapper — Type

Wraps SSP Dataset (CSV file).

Read GeoTIFF

Main.DIVACoast.geotiff_connect — Function

This function connect two geotiffs by a given operation. infilename1::String infilename2::String outfilename::String f::Function

Spatial Operations

SparseGeoArray (SGA)

Main.DIVACoast.SparseGeoArray — Type

Base.getindex — Function

getindex(sga::SparseGeoArray, i::AbstractRange, j::AbstractRange, k::Union{Colon,AbstractRange,Integer})

Index a SparseGeoArray with AbstractRanges to get a cropped SparseGeoArray with the correct AffineMap set.

Examples

julia> sga[2:3,2:3]
2x2x1 Array{Float64, 3} with AffineMap([1.0 0.0; 0.0 1.0], [1.0, 1.0]) and undefined CRS

Main.DIVACoast.coords — Function

coords(sga::SparseGeoArray, p::SVector{2,<:Integer}, strategy::AbstractStrategy=Center())
coords(sga::SparseGeoArray, p::Tuple{<:Integer,<:Integer}, strategy::AbstractStrategy=Center())
coords(sga::SparseGeoArray, p::CartesianIndex{2}, strategy::AbstractStrategy=Center())

Retrieve coordinates of the cell index by p. See indices for the inverse function.

Main.DIVACoast.indices — Function

indices(sga::SparseGeoArray, p::SVector{2,<:Real})

Retrieve logical indices of the cell represented by coordinates p. See coords for the inverse function.

Main.DIVACoast.nh4 — Function

nh4(sga :: SparseGeoArray{DT, IT}, x :: Integer, y :: Integer) :: Array{Tuple{IT,IT}}
nh4(sga :: SparseGeoArray{DT, IT}, p :: Tuple{Integer,Integer}) :: Array{Tuple{IT,IT}}

Compute the 4-Neighbourhood of the grid cell x,y in the SparseGeoArray sga and return as Array of pairs. Takes into account the boundaries of the SparseGeoArray.

Examples

julia> nh4(sga, 1, 1)
2-element Vector{Tuple{Int32, Int32}}:
 (2, 1)
 (1, 2)

julia> nh4(sga, 2, 4)
4-element Vector{Tuple{Int32, Int32}}:
 (1, 4)
 (3, 4)
 (2, 3)
 (2, 5)

Main.DIVACoast.nh8 — Function

same as nh4 but accounting for al 8 neighbours of a pixel with index (x,y)

Main.DIVACoast.distance — Function

distance(lon1 :: R, lat1 :: R, lon2 :: R, lat2 :: R) :: R where {R <: Real}
distance(p1 :: SVector{2,R}, p2 :: SVector{2,R}) :: R where {R <: Real}
distance(p1 :: AbstractVector{R}, p2 :: AbstractVector{R}) :: R where {R <: Real} 
distance(p1 :: Tuple{R,R}, p2 :: Tuple{R,R}) where {R <: Real} :: R where {R <: Real}

Compute the distance (in km) between two points given by lon1,lat1 and lon2,lat2 resp. p1 and p2. Uses the Haversine formula.

Examples

julia> ...

distance(hspf::HypsometricProfile, e::Real)

Compute the distance of elevation e (given in m) from the coastline in hspf. disatnce is returned in km.

Main.DIVACoast.go_direction — Function

go_direction(lon :: R, lat :: R, distance :: Real, direction :: AbstractDirection) :: Tuple{R,R} where {R <: Real}

Compute the geographical coordinates of the point reached if we go distance km from (lon,lat) in direction. Takes into account circularity, but does not cross poles. direction can be East(), North(), West(), South()

Examples

julia> go_direction(13.2240, 52.3057, 10, East())
(13.370916039175427, 52.3057)
julia> go_direction(13.2240, 52.3057, 10000, North())
(13.224, 90.0)
julia> go_direction(19.0045,0.0,40075,West())
(19.004500000000007, 0.0)

Main.DIVACoast.bounding_boxes — Function

bounding_boxes(sga :: SparseGeoArray{DT, IT}, lon_east :: Real, lon_west :: Real, lat_south :: Real, lat_north :: Real) where {DT <: Real, IT <: Integer}

Compute the bounding box(es) for the sparse geoarray sga and an area from loneast to lonwest and latsouth and latnorth.

Examples

julia> ...

Main.DIVACoast.area — Function

Main.DIVACoast.emptySGAfromSGA — Function

emptySGAfromSGA(orgSGA::SparseGeoArray{DT,IT}, extentNew)

Creates an empty SparseGeoArray from an existing SGA (orgSGA) with same projection / specifications.

Main.DIVACoast.get_extent — Function

Get the extent of an SparseGeoArray.

Main.DIVACoast.sga_union — Function

sga_union(sgaArray::Array{SparseGeoArray{DT,IT}}) Get the union of multiple SparseGeoArrays.

Main.DIVACoast.sga_intersect — Function

Get the intersetct of multiple SparseGeoArrays.

Main.DIVACoast.sga_diff — Function

Get the difference of multiple SparseGeoArrays.

Main.DIVACoast.sga_summarize_within — Function

Summarize data within a certain radius around a point p using a summarize function.

Main.DIVACoast.minumum_mean — Function

Function to get the mean of all minimum values according to sortlist in valuelist

Main.DIVACoast.get_closest_value — Function

Get the closest value to a point p. Should be used for small datasets only

Main.DIVACoast.get_box_around — Function

Crop a SGA to an extent defined by a radius around a point p.

Main.DIVACoast.epsg2wkt — Function

Get the WKT of an Integer EPSG code

Main.DIVACoast.proj2wkt — Function

Get the WKT (WellKnownText) of an Proj string

Main.DIVACoast.str2wkt — Function

Parse CRS string into WKT.

Main.DIVACoast.epsg! — Function

Set CRS on SparseGeoArray by epsgcode

Main.DIVACoast.is_rotated — Function

Check wether the AffineMap of a GeoArray contains rotations.

Main.DIVACoast.bbox! — Function

Set geotransform of SparseGeoArray by specifying a bounding box. Note that this only can result in a non-rotated or skewed GeoArray.

Spatial-Relationship

Main.DIVACoast.Neighbour — Type

The Neighbours structure holds and BallTree Object and the created Matrix.

Main.DIVACoast.nearest — Function

The nearest function returns the index of nearest neighbour of an Neighbours Object to an coordinate.

Parameter

n: The Neighbours Object to search trough.

coordinate: A coordinate the nearest neighbour relates to.

The nearest function returns the nearest neighbour of an Neighbours Object to multiple coordinates. Coordinates can be passed to the fucntion as a DataFrame.

Parameter

n: The Neighbours Object to search trough.

df: DataFrame holding the coordinates, the nearest neighbour should relate to.

dtype: DataType the coordinates will be parsed in.

lonlatCols: Columns names (string / symbol) of the columns in the input

dataframe holding the coordinates.

dropna: Whether na's should be kept or not.

Main.DIVACoast.nearest_coord — Function

Does same as nearest() but returns Coordinate of nearest neighbour.

Main.DIVACoast.coords_to_wide — Function

This funciton transforms the longitude and latitude columns of an DataFrame to a Matrix in wide-format required for Nearest Neighbour matching.

Parameter

df: The input DataFrame containing the longitude and latitude column.

dtype: The datatype the coordinates should be parsed in.

lonlatCols: The longitude and latitude columns (default = (:lon, :lat))

Return

returns created matrix and DataFrame the matrix is based on (omitted NA values).